Refactoring some packages: base package, main, greeting Added scheduling : temp placed to run every christmas lol
