+++ src/main/java/com/deocampo/gsb/Application.java
@@ -1,4 +1,4 @@
-package hello;
+package com.deocampo.gsb;
 
 import java.util.Arrays;
 
@@ -7,8 +7,10 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
+@EnableScheduling
 public class Application {
 
     public static void main(String[] args) {
