+++ src/main/java/com/deocampo/gsb/greeting/Greeting.java
@@ -1,4 +1,4 @@
-package hello;
+package com.deocampo.gsb.greeting;
 
 public class Greeting {
