+++ src/main/java/com/deocampo/gsb/greeting/GreetingController.java
@@ -1,5 +1,4 @@
-
-package hello;
+package com.deocampo.gsb.greeting;
 
 import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.web.bind.annotation.RequestMapping;
