+++ src/main/java/com/deocampo/gsb/schedule/ScheduledTask.java
@@ -1,0 +1,22 @@
+package com.deocampo.gsb.schedule;
+
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.stereotype.Component;
+
+@Component
+public class ScheduledTask {
+
+    private static final Logger log = LoggerFactory.getLogger(ScheduledTask.class);
+
+    private static final SimpleDateFormat dateFormat = new SimpleDateFormat("HH:mm:ss");
+
+    @Scheduled(cron = "0 0 0 25 12 ?")
+    public void reportCurrentTime() {
+        log.info("The time is now {}", dateFormat.format(new Date()));
+    }
+}
